Initial Commit
Renaming custom_comonents to yamaha_soundbar

Also, removed sources not supported on soundbar.
