+++ custom_components/yamaha_soundbar/client.pem
