+++ custom_components/yamaha_soundbar/manifest.json
@@ -1,9 +1,9 @@
 {
-  "domain": "linkplay",
-  "name": "Linkplay",
-  "version":"3.1.8",
-  "documentation": "https://github.com/nagyrobi/home-assistant-custom-components-linkplay",
-  "issue_tracker": "https://github.com/nagyrobi/home-assistant-custom-components-linkplay/issues",
+  "domain": "yamaha_soundbar",
+  "name": "Yamaha Soundbar",
+  "version":"3.1.9",
+  "documentation": "https://github.com/osk2/yamaha-soundbar",
+  "issue_tracker": "https://github.com/osk2/yamaha-soundbar",
   "after_dependencies": ["http", "tts", "media_source"],
   "config_flow": false,
   "iot_class": "local_polling",
