+++ hacs.json
@@ -1,5 +1,5 @@
 {
-  "name": "YAMAHA Soundbar",
+  "name": "Yamaha Soundbar",
   "content_in_root": false,
   "render_readme": false,
   "domains": ["media_player"],
